Component RBAC for different roles on user token
