+++ src/views/dashboard/Dashboard.js
@@ -50,6 +50,7 @@ import avatar3 from 'src/assets/images/avatars/3.jpg'
 import avatar4 from 'src/assets/images/avatars/4.jpg'
 import avatar5 from 'src/assets/images/avatars/5.jpg'
 import avatar6 from 'src/assets/images/avatars/6.jpg'
+import ComponentRBAC from 'src/utils/ComponentRBAC'
 
 const Dashboard = () => {
   const random = (min, max) => Math.floor(Math.random() * (max - min + 1) + min)
@@ -177,133 +178,134 @@ const Dashboard = () => {
 
   return (
     <>
-      <CCard className="mb-4">
-        <CCardBody>
-          <CRow>
-            <CCol sm={5}>
-              <h4 id="traffic" className="card-title mb-0">
-                Traffic
-              </h4>
-              <div className="small text-medium-emphasis">January - July 2021</div>
-            </CCol>
-            <CCol sm={7} className="d-none d-md-block">
-              <CButton color="primary" className="float-end">
-                <CIcon icon={cilCloudDownload} />
-              </CButton>
-              <CButtonGroup className="float-end me-3">
-                {['Day', 'Month', 'Year'].map((value) => (
-                  <CButton
-                    color="outline-secondary"
-                    key={value}
-                    className="mx-0"
-                    active={value === 'Month'}
-                  >
-                    {value}
-                  </CButton>
-                ))}
-              </CButtonGroup>
-            </CCol>
-          </CRow>
-          <CChartLine
-            style={{ height: '300px', marginTop: '40px' }}
-            data={{
-              labels: ['January', 'February', 'March', 'April', 'May', 'June', 'July'],
-              datasets: [
-                {
-                  label: 'My First dataset',
-                  backgroundColor: hexToRgba(getStyle('--cui-info'), 10),
-                  borderColor: getStyle('--cui-info'),
-                  pointHoverBackgroundColor: getStyle('--cui-info'),
-                  borderWidth: 2,
-                  data: [
-                    random(50, 200),
-                    random(50, 200),
-                    random(50, 200),
-                    random(50, 200),
-                    random(50, 200),
-                    random(50, 200),
-                    random(50, 200),
-                  ],
-                  fill: true,
-                },
-                {
-                  label: 'My Second dataset',
-                  backgroundColor: 'transparent',
-                  borderColor: getStyle('--cui-success'),
-                  pointHoverBackgroundColor: getStyle('--cui-success'),
-                  borderWidth: 2,
-                  data: [
-                    random(50, 200),
-                    random(50, 200),
-                    random(50, 200),
-                    random(50, 200),
-                    random(50, 200),
-                    random(50, 200),
-                    random(50, 200),
-                  ],
-                },
-                {
-                  label: 'My Third dataset',
-                  backgroundColor: 'transparent',
-                  borderColor: getStyle('--cui-danger'),
-                  pointHoverBackgroundColor: getStyle('--cui-danger'),
-                  borderWidth: 1,
-                  borderDash: [8, 5],
-                  data: [65, 65, 65, 65, 65, 65, 65],
-                },
-              ],
-            }}
-            options={{
-              maintainAspectRatio: false,
-              plugins: {
-                legend: {
-                  display: false,
-                },
-              },
-              scales: {
-                x: {
-                  grid: {
-                    drawOnChartArea: false,
-                  },
-                },
-                y: {
-                  ticks: {
-                    beginAtZero: true,
-                    maxTicksLimit: 5,
-                    stepSize: Math.ceil(250 / 5),
-                    max: 250,
-                  },
-                },
-              },
-              elements: {
-                line: {
-                  tension: 0.4,
-                },
-                point: {
-                  radius: 0,
-                  hitRadius: 10,
-                  hoverRadius: 4,
-                  hoverBorderWidth: 3,
-                },
-              },
-            }}
-          />
-        </CCardBody>
-        <CCardFooter>
-          <CRow xs={{ cols: 1 }} md={{ cols: 5 }} className="text-center">
-            {progressExample.map((item, index) => (
-              <CCol className="mb-sm-2 mb-0" key={index}>
-                <div className="text-medium-emphasis">{item.title}</div>
-                <strong>
-                  {item.value} ({item.percent}%)
-                </strong>
-                <CProgress thin className="mt-2" color={item.color} value={item.percent} />
+      <ComponentRBAC allowedRoles={['admin', 'member']}>
+        <CCard className="mb-4">
+          <CCardBody>
+            <CRow>
+              <CCol sm={5}>
+                <h4 id="traffic" className="card-title mb-0">
+                  Traffic
+                </h4>
+                <div className="small text-medium-emphasis">January - July 2021</div>
               </CCol>
-            ))}
-          </CRow>
-        </CCardFooter>
-      </CCard>
-
+              <CCol sm={7} className="d-none d-md-block">
+                <CButton color="primary" className="float-end">
+                  <CIcon icon={cilCloudDownload} />
+                </CButton>
+                <CButtonGroup className="float-end me-3">
+                  {['Day', 'Month', 'Year'].map((value) => (
+                    <CButton
+                      color="outline-secondary"
+                      key={value}
+                      className="mx-0"
+                      active={value === 'Month'}
+                    >
+                      {value}
+                    </CButton>
+                  ))}
+                </CButtonGroup>
+              </CCol>
+            </CRow>
+            <CChartLine
+              style={{ height: '300px', marginTop: '40px' }}
+              data={{
+                labels: ['January', 'February', 'March', 'April', 'May', 'June', 'July'],
+                datasets: [
+                  {
+                    label: 'My First dataset',
+                    backgroundColor: hexToRgba(getStyle('--cui-info'), 10),
+                    borderColor: getStyle('--cui-info'),
+                    pointHoverBackgroundColor: getStyle('--cui-info'),
+                    borderWidth: 2,
+                    data: [
+                      random(50, 200),
+                      random(50, 200),
+                      random(50, 200),
+                      random(50, 200),
+                      random(50, 200),
+                      random(50, 200),
+                      random(50, 200),
+                    ],
+                    fill: true,
+                  },
+                  {
+                    label: 'My Second dataset',
+                    backgroundColor: 'transparent',
+                    borderColor: getStyle('--cui-success'),
+                    pointHoverBackgroundColor: getStyle('--cui-success'),
+                    borderWidth: 2,
+                    data: [
+                      random(50, 200),
+                      random(50, 200),
+                      random(50, 200),
+                      random(50, 200),
+                      random(50, 200),
+                      random(50, 200),
+                      random(50, 200),
+                    ],
+                  },
+                  {
+                    label: 'My Third dataset',
+                    backgroundColor: 'transparent',
+                    borderColor: getStyle('--cui-danger'),
+                    pointHoverBackgroundColor: getStyle('--cui-danger'),
+                    borderWidth: 1,
+                    borderDash: [8, 5],
+                    data: [65, 65, 65, 65, 65, 65, 65],
+                  },
+                ],
+              }}
+              options={{
+                maintainAspectRatio: false,
+                plugins: {
+                  legend: {
+                    display: false,
+                  },
+                },
+                scales: {
+                  x: {
+                    grid: {
+                      drawOnChartArea: false,
+                    },
+                  },
+                  y: {
+                    ticks: {
+                      beginAtZero: true,
+                      maxTicksLimit: 5,
+                      stepSize: Math.ceil(250 / 5),
+                      max: 250,
+                    },
+                  },
+                },
+                elements: {
+                  line: {
+                    tension: 0.4,
+                  },
+                  point: {
+                    radius: 0,
+                    hitRadius: 10,
+                    hoverRadius: 4,
+                    hoverBorderWidth: 3,
+                  },
+                },
+              }}
+            />
+          </CCardBody>
+          <CCardFooter>
+            <CRow xs={{ cols: 1 }} md={{ cols: 5 }} className="text-center">
+              {progressExample.map((item, index) => (
+                <CCol className="mb-sm-2 mb-0" key={index}>
+                  <div className="text-medium-emphasis">{item.title}</div>
+                  <strong>
+                    {item.value} ({item.percent}%)
+                  </strong>
+                  <CProgress thin className="mt-2" color={item.color} value={item.percent} />
+                </CCol>
+              ))}
+            </CRow>
+          </CCardFooter>
+        </CCard>
+      </ComponentRBAC>
       <CRow>
         <CCol xs>
           <CCard className="mb-4">
